+++ src/asset_vfx.h
@@ -2,36 +2,13 @@
 #define __ASS_PART__
 #include "game_define.h"
 #include "asset_define.h"
+#include "vfx_enum.h"
 
 #define MAX_PARTICLES        256
 #define MAX_PARTICLE_LAYERS  LAYER_DONE
 
 Entity ParticlePoolGetAvailable(RenderLayer l);
 
-DEFINE_EVENT_SPACE(ParticleEvent, EVENT_PARTICLE_BASE);
-
-typedef enum {
-  PARTICLE_EVENT_BURST,
-  PARTICLE_EVENT_START,
-  PARTICLE_EVENT_STOP,
-  PARTICLE_EVENT_COUNT
-} ParticleEventID;
-
-typedef enum {
-  PARTICLE_NONE,
-  PARTICLE_SPRITE,
-  PARTICLE_RECT,
-  PARTICLE_CIRCLE,
-  PARTICLE_PIXEL,
-} ParticleDrawType;
-
-typedef enum {
-  PARTICLE_EMIT_NONE,
-  PARTICLE_EMIT_CONTINUOUS,
-  PARTICLE_EMIT_BURST,
-  PARTICLE_EMIT_EVENT
-} ParticleEmitMode;
-  
 typedef struct {
   bool             active;
   ParticleDrawType draw_type;
