+++ src/process_enum.h
@@ -1,0 +1,36 @@
+#ifndef __PROC_ENUM__
+#define __PROC_ENUM__
+
+DEFINE_EVENT_SPACE(GameEvent, EVENT_GAME_BASE)
+
+typedef enum{
+  PROCESS_NONE,
+  PROCESS_LEVEL,
+  PROCESS_DONE
+}ProcessType;
+
+typedef enum{
+  UPDATE_FRAME,//update running at full fps
+  UPDATE_DRAW_BEGIN,
+  UPDATE_DRAW,
+  UPDATE_DRAW_END,
+  UPDATE_PRE,
+  UPDATE_FIXED,
+  UPDATE_POST,
+  UPDATE_FINAL,
+  UPDATE_DONE
+}UpdateType;
+
+typedef enum{
+  GAME_NONE,
+  GAME_LOADING,
+  GAME_READY,
+  GAME_RUNNING,
+  GAME_PAUSE,
+  GAME_FINISHED,
+  GAME_OVER,
+  GAME_DONE
+}GameState;
+
+
+#endif
